refactor: Divide local storage controller to store class and search history store class
